// src/app/interfaces/categoria-producto.enum.ts
export enum CategoriaProducto {
  BEBIDA = 'bebida',
  CARNE = 'carne',
  CONDIMENTO = 'condimento',
  CONGELADOS = 'congelados',
  DESAYUNO = 'desayuno',
  DULCES = 'dulces',
  FIAMBRES = 'fiambres',
  FRUTAS = 'frutas',
  HIGIENE = 'higiene',
  LACTEO = 'lacteo',
  LEGUMBRES = 'legumbres',
  LIMPIEZA = 'limpieza',
  MASCOTAS = 'mascotas',
  OTROS = 'otros',
  PANADERIA = 'panaderia',
  PASTAS = 'pastas',
  SNACK = 'snack',
  VERDURA = 'verdura'
}